+++ test/harnesses/LinTokenBasedFormulaHarness.sol
@@ -8,18 +8,17 @@ contract LinTokenBasedFormulaHarness is LinTokenBasedFormula {
     function calculateLinTokenBasedPurchaseCost(
         uint256 supply,
         uint256 initialCost,
-        uint256 scalingFactor,
+        uint32 scalingFactor,
         uint256 amount
     ) public pure returns (uint256 totalCost) {
         totalCost = _calculateLinTokenBasedPurchaseCost(supply, initialCost, scalingFactor, amount);
     }
 
-    function calculateLinTokenBasedSaleReturn(
-        uint256 supply,
-        uint256 initialCost,
-        uint256 scalingFactor,
-        uint256 amount
-    ) public pure returns (uint256 saleReturn) {
+    function calculateLinTokenBasedSaleReturn(uint256 supply, uint256 initialCost, uint32 scalingFactor, uint256 amount)
+        public
+        pure
+        returns (uint256 saleReturn)
+    {
         saleReturn = _calculateLinTokenBasedSaleReturn(supply, initialCost, scalingFactor, amount);
     }
 }
